+++ src/mixins/dateUtil.js
@@ -1,16 +1,30 @@
 import { pad } from '@/util';
 
+export function prettyDate(date) {
+  const tmp = date != null ? date : new Date();
+  const year = tmp.getFullYear();
+  const month = pad(tmp.getMonth() + 1);
+  const day = pad(tmp.getDate());
+  const hour = pad(tmp.getHours());
+  const minute = pad(tmp.getMinutes());
+  const second = pad(tmp.getSeconds());
+  return `${year}-${month}-${day} ${hour}:${minute}:${second}`;
+}
+
+export function safeDate(date) {
+  const tmp = date != null ? date : new Date();
+  const year = tmp.getFullYear();
+  const month = pad(tmp.getMonth() + 1);
+  const day = pad(tmp.getDate());
+  const hour = pad(tmp.getHours());
+  const minute = pad(tmp.getMinutes());
+  const second = pad(tmp.getSeconds());
+  return `${year}-${month}-${day}_${hour}-${minute}-${second}`;
+}
+
 export default {
   filters: {
-    prettyDate(date) {
-      if (!date) return null;
-      const year = date.getFullYear();
-      const month = pad(date.getMonth() + 1);
-      const day = pad(date.getDate());
-      const hour = pad(date.getHours());
-      const minute = pad(date.getMinutes());
-      const second = pad(date.getSeconds());
-      return `${year}-${month}-${day} ${hour}:${minute}:${second}`;
-    },
+    prettyDate,
+    safeDate,
   },
 };